Implement registration flow, add faculty management (add/delete), and finalize UI components

+++ app/src/main/java/com/example/universityapp/data/local/FacultyEntity.kt
@@ -4,7 +4,7 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity(tableName = "faculties")
-data class FacultyEntity(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
+data class  FacultyEntity(
+    @PrimaryKey(autoGenerate = true) val id: Long    = 0,
     val name: String
 )
